+++ mithril-common/src/messages/mod.rs
@@ -1,6 +1,7 @@
 //! Messages module
 //! This module aims at providing shared structures for API communications.
 mod aggregator_features;
+mod aggregator_status;
 mod cardano_stake_distribution;
 mod cardano_stake_distribution_list;
 mod cardano_transaction_snapshot;
@@ -23,6 +24,7 @@ mod snapshot_list;
 pub use aggregator_features::{
     AggregatorCapabilities, AggregatorFeaturesMessage, CardanoTransactionsProverCapabilities,
 };
+pub use aggregator_status::AggregatorStatusMessage;
 pub use cardano_stake_distribution::CardanoStakeDistributionMessage;
 pub use cardano_stake_distribution_list::{
     CardanoStakeDistributionListItemMessage, CardanoStakeDistributionListMessage,